requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.4
Deprecated==1.2.14
Django==4.2.5
django-extensions==3.2.3
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
djangorestframework-oauth==1.1.0
idna==3.4
jsonmodels==2.6.0
jwcrypto==1.5.0
MarkupSafe==2.1.3
oauthlib==3.2.2
Pillow==10.0.1
psycopg2-binary==2.9.9
pycparser==2.21
pyOpenSSL==23.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.6
Werkzeug==3.0.0
whitenoise==6.6.0
wrapt==1.15.0
